Refactored the project to use the new namespace and more
Also made some small stylistic modifications and wrote out a roadmap for future expansion
