+++ README.md
@@ -1,3 +1,11 @@
 # Simple Quern
 
-Simple Quern is a mod for Minecraft v1.15.x that adds a simple version of ore doubling in the form of the quern.
+Simple Quern is a mod for Minecraft 1.16.2 that adds a simple version of ore doubling in the form of the quern.
+
+### Roadmap
+- [ ] Replace LibCD with https://github.com/FabricMC/fabric/pull/1075
+- [ ] Turn the quern into a block entity renderer with rotation animation on use
+    - [ ] Allow the handstone in the block to correspond with the handstone in the quern
+- [ ] Rewrite the recipe deserializer with Codecs
+    - [ ] Rewrite the recipes to use loot tables
+- [ ] Allow the handstones to be enchanted with efficiency and fortune
