+++ gradle.properties
@@ -8,8 +8,8 @@ org.gradle.jvmargs=-Xmx1G
 	loader_version=0.9.1+build.205
 
 # Mod Properties
-	mod_version = 1.1.1
-	maven_group = com.wtoll
+	mod_version = 1.2.0
+	maven_group = com.cumulus
 	archives_base_name = simplequern
 
 # Dependencies
