+++ src/main/java/com/cumulus/simplequern/client/SimpleQuernClient.java
@@ -1,12 +1,11 @@
-package com.wtoll.simplequern.client;
+package com.cumulus.simplequern.client;
 
-import com.wtoll.simplequern.screen.QuernScreen;
-import com.wtoll.simplequern.screen.ScreenHandlers;
+import com.cumulus.simplequern.screen.QuernScreen;
+import com.cumulus.simplequern.screen.ScreenHandlers;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.screenhandler.v1.ScreenRegistry;
 
-public class SimpleQuernClient implements ClientModInitializer
-{
+public class SimpleQuernClient implements ClientModInitializer {
 	@Override
 	public void onInitializeClient() {
 		ScreenRegistry.register(ScreenHandlers.QUERN, QuernScreen::new);
