+++ src/main/java/com/cumulus/simplequern/recipe/RecipeSerializer.java
@@ -1,6 +1,6 @@
-package com.wtoll.simplequern.recipe;
+package com.cumulus.simplequern.recipe;
 
-import com.wtoll.simplequern.Utility;
+import com.cumulus.simplequern.SimpleQuern;
 import net.minecraft.recipe.Recipe;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
@@ -9,7 +9,7 @@ public class RecipeSerializer {
     public static final net.minecraft.recipe.RecipeSerializer<GrindingRecipe> GRINDING;
 
     static {
-        GRINDING = register(Utility.id("grinding"), new GrindingRecipeSerializer(GrindingRecipe::new));
+        GRINDING = register(SimpleQuern.id("grinding"), new GrindingRecipeSerializer(GrindingRecipe::new));
     }
 
     static <S extends net.minecraft.recipe.RecipeSerializer<T>, T extends Recipe<?>> S register(Identifier id, S serializer) {
