+++ src/main/java/com/cumulus/simplequern/recipe/RecipeType.java
@@ -1,6 +1,6 @@
-package com.wtoll.simplequern.recipe;
+package com.cumulus.simplequern.recipe;
 
-import com.wtoll.simplequern.Utility;
+import com.cumulus.simplequern.SimpleQuern;
 import net.minecraft.recipe.Recipe;
 import net.minecraft.util.registry.Registry;
 
@@ -12,7 +12,7 @@ public class RecipeType {
     }
 
     public static <T extends Recipe<?>> net.minecraft.recipe.RecipeType<T> register(final String string) {
-        return Registry.register(Registry.RECIPE_TYPE, Utility.id(string), new net.minecraft.recipe.RecipeType<T>() {
+        return Registry.register(Registry.RECIPE_TYPE, SimpleQuern.id(string), new net.minecraft.recipe.RecipeType<T>() {
             public String toString() {
                 return string;
             }
