--- src/main/java/com/wtoll/simplequern/Utility.java
@@ -1,9 +1,0 @@
-package com.wtoll.simplequern;
-
-import net.minecraft.util.Identifier;
-
-public class Utility {
-    public static Identifier id(String s) {
-        return new Identifier(SimpleQuern.MODID, s);
-    }
-}
